+++ test/hmac_md5_test.dart
@@ -11,7 +11,7 @@ void main() {
   group("standard vector", () {
     for (var i = 0; i < _inputs.length; i++) {
       test(_macs[i], () {
-        expectHmacEquals(new MD5(), _inputs[i], _keys[i], _macs[i]);
+        expectHmacEquals(md5, _inputs[i], _keys[i], _macs[i]);
       });
     }
   });