+++ test/hmac_sha256_test.dart
@@ -12,7 +12,7 @@ void main() {
   group("standard vector", () {
     for (var i = 0; i < _inputs.length; i++) {
       test(_macs[i], () {
-        expectHmacEquals(new SHA256(), _inputs[i], _keys[i], _macs[i]);
+        expectHmacEquals(sha256, _inputs[i], _keys[i], _macs[i]);
       });
     }
   });